requirements.txt
Django
django-notify-x
django-bootstrap3
django-braces
